+++ app/shared/components/design-system/all-components/navigation-bar/mobile-nav/mobile-overlay/MobileOverlay.styles.ts
