+++ app/shared/components/design-system/all-components/navigation-bar/mobile-nav/mobile-overlay/nav-contacts/NavContactsSection.styles.ts
@@ -1,6 +1,5 @@
-import { mainHexPallete } from '../../../theme/colors';
-
 import { AppTypography } from '~/constants';
+import { mainHexPallete } from '~/shared/components/design-system/all-components/theme/colors';
 
 export const styles = {
   links: {
